+++ src/OdfAttributeName.ts
@@ -5,11 +5,12 @@ export enum OdfAttributeName {
   FormatFontStyle = "fo:font-style",
   FormatFontWeight = "fo:font-weight",
   FormatTextAlign = "fo:text-align",
-
+  
   OfficeMimetype = "office:mimetype",
   OfficeVersion = "office:version",
-
+  
   StyleFamily = "style:family",
+  StyleFontName = "style:font-name",
   StyleName = "style:name",
   StylePosition = "style:position",
   StyleType = "style:type",