+++ src/style/FontPitch.ts
@@ -1,0 +1,4 @@
+export enum FontPitch {
+  Fixed = "fixed",
+  Variable = "variable",
+}